Kumar dev

+++ resources/js/Pages/Device/Device.vue
@@ -2,23 +2,34 @@
 import AuthenticatedLayout from "@/Layouts/AuthenticatedLayout.vue";
 import { Head } from "@inertiajs/vue3";
 import { Inertia } from "@inertiajs/inertia";
+import { ref } from "vue";
 import { toast } from "vue3-toastify";
 import "vue3-toastify/dist/index.css";
 import axios from "axios";
 
+
+const loadingButtons = ref({});
+
 // Delete device
 const deleteDevice = async (id) => {
-  try {
-    const response = await axios.delete(`devices/destroy/${id}`);
-    toast.success(response.data.message);
-    // Reload the page
-    setTimeout(function () {
-      window.location.reload();
-    }, 3000);
-  } catch (error) {
-    const errorMessage =
-      error.response?.data?.message || "An error occurred. Please try again.";
-    toast.error(errorMessage);
+  if (confirm("Are you sure you want to delete this device?")) {
+    loadingButtons.value[id] = true;
+    try {
+      const response = await axios.delete(`devices/destroy/${id}`);
+      toast.success(response.data.message);
+      // Reload the page
+      setTimeout(function () {
+        window.location.reload();
+      }, 3000);
+    } catch (error) {
+      const errorMessage =
+        error.response?.data?.message || "An error occurred. Please try again.";
+      toast.error(errorMessage);
+    }
+    setTimeout(() => {
+      loadingButtons.value[id] = false; // Reset the loading state
+      window.location.reload(); // Reload the page
+    }, 4000);
   }
 };
 </script>
@@ -100,11 +111,21 @@ const deleteDevice = async (id) => {
 
                 <!-- Delete button -->
                 <button
-                  class="btn btn-danger action-btn"
-                  @click="deleteDevice(device.id)"
-                >
-                  <i class="bi bi-trash-fill"></i>
-                </button>
+                    class="btn btn-danger action-btn"
+                    :disabled="loadingButtons[device.id]"
+                    @click="deleteDevice(device.id)"
+                  >
+                    <!-- Show trash icon if not loading -->
+                    <span v-if="!loadingButtons[device.id]">
+                      <i class="bi bi-trash-fill"></i>
+                    </span>
+                    
+                    <!-- Show spinner and Deleting text if loading -->
+                    <span v-else>
+                      <i class="spinner-border spinner-border-sm" role="status"></i>
+                      Deleting...
+                    </span>
+                  </button>
               </td>
             </tr>
           </tbody>
